Update WirelessTagsMotion.groovy
commented out arm and rest call in the update. It was causing the init to tie out and improperly resetting closed states whenever you add a new tag or change the config in the app.

I also added the sensor capability so it will show up in the action tiles list.